requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
python-docx
reportlab
pillow
